관람평 구현 완료

+++ src/main/java/com/bclass/arts_center/dto/request/RequestReviewDto.java
@@ -1,11 +1,15 @@
 package com.bclass.arts_center.dto.request;
 
 import lombok.Data;
-
+/**
+ * 
+ * @author 전대영 
+ *
+ */
 @Data
 public class RequestReviewDto {
 	private Integer show_id;
-	private Integer rating;
+	private Double rating;
 	private String title;
 	private String content;
 	private String reviewCreationDate;
